- windows x64 updates to get things compiling on vc2012 x64

+++ source/glest_game/ai/ai_interface.cpp
@@ -833,7 +833,7 @@ bool AiInterface::isFreeCells(const Vec2i &pos, int size, Field field){
 }
 
 void AiInterface::removeEnemyWarningPositionFromList(Vec2i &checkPos) {
-	for(int i = enemyWarningPositionList.size() - 1; i >= 0; --i) {
+	for(int i = (int)enemyWarningPositionList.size() - 1; i >= 0; --i) {
 		Vec2i &pos = enemyWarningPositionList[i];
 
 		if(checkPos == pos) {
